backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.118.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
joblib==1.5.2
llvmlite==0.45.1
multidict==6.7.0
numba==0.62.1
numpy==2.2.6
openai==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
postgrest==2.22.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
pynndescent==0.5.13
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.22.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.9.post2
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
yt-dlp==2024.8.6
gallery-dl==1.27.0
